cleanup example. add deprecated notices
