+++ geometries_sphere_geometry.go
@@ -1,31 +1,56 @@
 package three
 
-//go:generate go run geometry_method_generator/main.go -geometryType CircleGeometry -geometrySlug circle_geometry
+//go:generate go run geometry_method_generator/main.go -geometryType SphereGeometry -geometrySlug sphere_geometry
 
-import "github.com/gopherjs/gopherjs/js"
+import (
+	"math"
 
-type CircleGeometry struct {
+	"github.com/gopherjs/gopherjs/js"
+)
+
+type SphereGeometry struct {
 	*js.Object
 
-	Radius      float64 `js:"radius"`
-	Segments    float64 `js:"segments"`
+	Radius float64 `js:"radius"`
+	// Segments    int     `js:"segments"` // Not sure if available.
 	ThetaStart  float64 `js:"thetaStart"`
 	ThetaLength float64 `js:"thetaLength"`
 }
 
-type CircleGeometryParameters struct {
-	Radius      float64
-	Segments    float64
-	ThetaStart  float64
-	ThetaLength float64
+type SphereGeometryParameters struct {
+	Radius         float64
+	WidthSegments  int
+	HeightSegments int
+	ThetaStart     float64
+	ThetaLength    float64
+	PhiStart       float64
+	PhiLength      float64
 }
 
-// NewBoxGeometry creates a new BoxGeometry.
-func NewCircleGeometry(params CircleGeometryParameters) CircleGeometry {
-	return CircleGeometry{
-		Object: three.Get("CircleGeometry").New(
+// NewSphereGeometry creates a new BoxGeometry.
+func NewSphereGeometry(params SphereGeometryParameters) SphereGeometry {
+	if params.ThetaLength == 0 {
+		params.ThetaLength = math.Pi
+	}
+	if params.PhiLength == 0 {
+		params.PhiLength = 2 * math.Pi
+	}
+	if params.WidthSegments == 0 {
+		params.WidthSegments = 8
+	}
+	if params.HeightSegments == 0 {
+		params.HeightSegments = 6
+	}
+	if params.Radius == 0 {
+		params.Radius = 1
+	}
+	return SphereGeometry{
+		Object: three.Get("SphereGeometry").New(
 			params.Radius,
-			params.Segments,
+			params.WidthSegments,
+			params.HeightSegments,
+			params.PhiStart,
+			params.PhiLength,
 			params.ThetaStart,
 			params.ThetaLength,
 		),
